refactor: Refactor symbol indexing logic and update import statement
