+++ vscode/bds/server/src/symbolIndex.ts
